add more props and update constant

+++ backend/schemas.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
 class EarningsResponse(BaseModel):
-    today_earnings: int
-    since_nov1_2025: int
-    total_earnings: int
+    today_earnings: float
+    since_nov1_2025: float
+    total_earnings: float
